refactor(#2827): split the probe's runtime gate + window policy into config.rs
partition_access/mod.rs crossed the campsite threshold with D3's env knobs, so the
gate (PROBE/WINDOW env parsing, enabled/set_probe_enabled) and WindowConfig move to
config.rs, re-exported unchanged. mod.rs 816 -> 662.
